@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 module Actions
   module ForemanPulsible
     module Pulp3
@@ -6,7 +7,6 @@ module Actions
         module Remote
           module Role
             class Create < ::Actions::ForemanPulsible::Base::PulsibleAction
-
               input_format do
                 param :name, String, required: true
                 param :url, String, required: true
@@ -17,10 +17,12 @@ module Actions
               end
 
               def run
-                role_remote = PulpAnsibleClient::AnsibleRoleRemote.new({
-                                                                                     :name => input[:name],
-                                                                                     :url => input[:url],
-                                                                                   })
+                role_remote = PulpAnsibleClient::AnsibleRoleRemote.new(
+                  {
+                    name: input[:name],
+                    url: input[:url],
+                  }
+                )
                 response = ::ForemanPulsible::Pulp3::Ansible::Remote::Role::Create.new(role_remote).request
                 output.update(role_remote_create_response: response)
               end
