@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Actions
   module ForemanPulsible
     module Pulp3
@@ -5,7 +7,6 @@ module Actions
         module Content
           module Collection
             class List < ::Actions::ForemanPulsible::Base::PulsibleAction
-
               input_format do
                 param :repository_version_href, String, required: true
               end
@@ -15,10 +16,11 @@ module Actions
               end
 
               def run
-                response = ::ForemanPulsible::Pulp3::Ansible::Content::Collection::List.new(input[:repository_version_href]).request
+                response = ::ForemanPulsible::Pulp3::Ansible::Content::Collection::List.new(
+                  input[:repository_version_href]
+                ).request
                 output.update(repository_artifacts: response)
               end
-
             end
           end
         end
