@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 module Actions
   module ForemanPulsible
     module Pulp3
@@ -17,7 +18,9 @@ module Actions
               end
 
               def invoke_external_task
-                response = ::ForemanPulsible::Pulp3::Ansible::Remote::Collection::Destroy.new(input[:collection_remote_href]).request
+                response = ::ForemanPulsible::Pulp3::Ansible::Remote::Collection::Destroy.new(
+                  input[:collection_remote_href]
+                ).request
                 output.update(collection_remote_destroy_response: response)
                 nil
               end
@@ -28,10 +31,10 @@ module Actions
 
               def poll_external_task
                 collection_remote_destroy_task = output&.[](:collection_remote_destroy_response)&.[](:task)
-                t = ::ForemanPulsible::Pulp3::Core::Task::Status.new(collection_remote_destroy_task).request
-                t = ::Parsers::Pulp3::Core::Task::Status.new(t)
+                task = ::ForemanPulsible::Pulp3::Core::Task::Status.new(collection_remote_destroy_task).request
+                task_status = ::Parsers::Pulp3::Core::Task::Status.new(task)
 
-                {progress: t.progress}
+                { progress: task_status.progress }
               end
 
               def task_output
