@@ -1,63 +1,63 @@
-module AnsibleContent
+# frozen_string_literal: true
 
+module AnsibleContent
   # Simple representation of an Ansible content unit
   # To be used in cases where AnsibleContentUnit is too heavy, i.e. Actions
   class SimpleAnsibleContentUnit
-    attr_reader :unit_type, :name, :versions, :source, :type, :src, :scm
-    attr_writer :versions
+    attr_accessor :versions
+    attr_reader :unit_type, :name, :source, :type, :src, :scm
 
     def initialize(**kwargs)
       @unit_type = kwargs[:unit_type]
       @name = kwargs[:unit_name]
       @versions = kwargs[:unit_versions] || []
-      @source = kwargs[:unit_source] || "https://galaxy.ansible.com/" # TODO: global param
+      @source = kwargs[:unit_source] || 'https://galaxy.ansible.com/' # TODO: global param
       @type = kwargs[:unit_type]
       @src = kwargs[:unit_src]
       @scm = kwargs[:unit_scm]
     end
 
+    # Name of the content unit with respect to the namespace:
+    # "theforeman.operations" --> "operations"
     def unit_name
-      @name.split(".").last
+      @name.split('.').last
     end
 
     # Namespace of the content unit
     # "theforeman.operations" --> "theforeman"
     def unit_namespace
-      @name.split(".").first
+      @name.split('.').first
     end
 
     def collection_file
-      if @unit_type == :collection
-        if @versions.length > 0
-          YAML.dump(
-            { "collections" => @versions.map { |version|
-              {
-                "name" => @name,
-                "version" => version,
-                "source" => @source,
-                "type" => @type
-              }.compact
-            }
-            })
-        else
-          YAML.dump(
-            { "collections" =>
-                [
-                  {
-                    "name" => @name,
-                    "source" => @source,
-                  }
-                ]
-            }
-          )
-        end
+      return unless @unit_type == :collection
+      if @versions.length.positive?
+        YAML.dump(
+          { 'collections' => @versions.map do |version|
+            {
+              'name' => @name,
+              'version' => version,
+              'source' => @source,
+              'type' => @type,
+            }.compact
+          end }
+        )
+      else
+        YAML.dump(
+          { 'collections' =>
+              [
+                {
+                  'name' => @name,
+                  'source' => @source,
+                },
+              ] }
+        )
       end
     end
 
     def role_url
-      if @unit_type == :role
-        "#{self.source}api/v1/roles/?namespace=#{self.unit_namespace}&name=#{self.unit_name}" # TODO: version restriction
-      end
+      return unless @unit_type == :role
+      "#{source}api/v1/roles/?namespace=#{unit_namespace}&name=#{unit_name}" # TODO: version restriction
     end
 
     def to_hash
@@ -68,9 +68,8 @@ module AnsibleContent
         source: @source,
         type: @type,
         src: @src,
-        scm: @scm
+        scm: @scm,
       }
     end
-
   end
 end
