@@ -1,26 +1,26 @@
+# frozen_string_literal: true
+
 module Parsers
   module Pulp3
     module Core
       module Task
         class Status < ::Parsers::Pulp3::ApiResponse
-
-          WAITING = 'waiting'.freeze
-          SKIPPED = 'skipped'.freeze
-          RUNNING = 'running'.freeze
-          COMPLETED = 'completed'.freeze
-          FAILED = 'failed'.freeze
-          CANCELED = 'canceled'.freeze
+          WAITING = 'waiting'
+          SKIPPED = 'skipped'
+          RUNNING = 'running'
+          COMPLETED = 'completed'
+          FAILED = 'failed'
+          CANCELED = 'canceled'
           FINISHED_STATES = [COMPLETED, FAILED, CANCELED, SKIPPED].freeze
 
           def task_completed?
-            FINISHED_STATES.include? @raw_response.dig('state')
+            FINISHED_STATES.include? @raw_response['state']
           end
 
           def progress
             return 1 if task_completed?
             0
           end
-
         end
       end
     end
