@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 class AnsibleRole < AnsibleContentUnit
-    has_many :ansible_content_versions, as: :versionable, dependent: :destroy
+  has_many :ansible_content_versions, as: :versionable, dependent: :destroy
 end
