@@ -1,13 +1,15 @@
+# frozen_string_literal: true
+
 module ForemanPulsible
   module Pulp3
     module Ansible
       class AnsibleApi
-        def initialize(*args)
+        def initialize(*_args)
           @ansible_api_client = Pulp3::BaseClient.ansible_api_client
         end
 
         def request
-          fail NotImplementedError
+          raise NotImplementedError
         end
       end
     end
