@@ -1,18 +1,20 @@
+# frozen_string_literal: true
+
 module ForemanPulsible
   module Pulp3
     module Ansible
       module Remote
         module Git
           class Destroy < GitRemoteApi
-          def initialize(git_remote)
-            super
-            @git_remote = git_remote
-          end
+            def initialize(git_remote)
+              super
+              @git_remote = git_remote
+            end
 
-          def request
-            @ansible_git_remote_api_client.delete(@git_remote)
+            def request
+              @ansible_git_remote_api_client.delete(@git_remote)
+            end
           end
-        end
         end
       end
     end
