@@ -1,6 +1,8 @@
+# frozen_string_literal: true
+
 module ForemanPulsible
   module Constants
     PLUGIN_NAME = 'foreman_pulsible'
-    PLUGIN_VERSION = '0.1.0'.freeze
+    PLUGIN_VERSION = '0.1.0'
   end
 end
